midterm-faster-rcnn/requirements.txt
scipy
numpy
matplotlib
opencv_python
torch
torchvision
tqdm
Pillow
h5py
torchsummary